Migrate getMemberInfo into a component
