+++ src/api/IBMiContent.ts
@@ -639,45 +639,6 @@ export default class IBMiContent {
   }
 
   /**
-   *
-   * @param filter: the criterias used to list the members
-   * @returns
-   */
-  async getMemberInfo(library: string, sourceFile: string, member: string): Promise<IBMiMember | undefined> {
-    if (this.ibmi.remoteFeatures[`GETMBRINFO.SQL`]) {
-      const tempLib = this.config.tempLibrary;
-      const statement = `select * from table(${tempLib}.GETMBRINFO('${library}', '${sourceFile}', '${member}'))`;
-
-      let results: Tools.DB2Row[] = [];
-      if (this.config.enableSQL) {
-        try {
-          results = await this.runSQL(statement);
-        } catch (e) {}; // Ignore errors, will return undefined.
-      }
-      else {
-        results = await this.getQTempTable([`create table QTEMP.MEMBERINFO as (${statement}) with data`], "MEMBERINFO");
-      }
-
-      if (results.length === 1 && results[0].ISSOURCE === 'Y') {
-        const result = results[0];
-        const asp = this.ibmi.aspInfo[Number(results[0].ASP)];
-        return {
-          library: result.LIBRARY,
-          file: result.FILE,
-          name: result.MEMBER,
-          extension: result.EXTENSION,
-          text: result.DESCRIPTION,
-          created: new Date(result.CREATED ? Number(result.CREATED) : 0),
-          changed: new Date(result.CHANGED ? Number(result.CHANGED) : 0)
-        } as IBMiMember
-      }
-      else {
-        return undefined;
-      }
-    }
-  }
-
-  /**
    * Get list of items in a path
    * @param remotePath
    * @return an array of IFSFile
