+++ src/testing/content.ts
@@ -7,6 +7,7 @@ import { Tools } from "../api/Tools";
 import { getMemberUri } from "../filesystems/qsys/QSysFs";
 import { instance } from "../instantiate";
 import { CommandResult } from "../typings";
+import { GetMemberInfo } from "../components/getMemberInfo";
 
 export const ContentSuite: TestSuite = {
   name: `Content API tests`,
@@ -606,8 +607,12 @@ export const ContentSuite: TestSuite = {
     {
       name: `Check getMemberInfo`, test: async () => {
         const content = instance.getContent();
+        const connection = instance.getConnection();
+        const component = connection?.getComponent<GetMemberInfo>(`GetMemberInfo`)!;
 
-        const memberInfoA = await content?.getMemberInfo(`QSYSINC`, `H`, `MATH` );
+        assert.ok(component);
+
+        const memberInfoA = await component.getMemberInfo(`QSYSINC`, `H`, `MATH` );
         assert.ok(memberInfoA);
         assert.strictEqual(memberInfoA?.library === `QSYSINC`, true);
         assert.strictEqual(memberInfoA?.file === `H`, true);
@@ -615,7 +620,7 @@ export const ContentSuite: TestSuite = {
         assert.strictEqual(memberInfoA?.extension === `C`, true);
         assert.strictEqual(memberInfoA?.text === `STANDARD HEADER FILE MATH`, true);
 
-        const memberInfoB = await content?.getMemberInfo(`QSYSINC`, `H`, `MEMORY` );
+        const memberInfoB = await component.getMemberInfo(`QSYSINC`, `H`, `MEMORY` );
         assert.ok(memberInfoB);
         assert.strictEqual(memberInfoB?.library === `QSYSINC`, true);
         assert.strictEqual(memberInfoB?.file === `H`, true);
@@ -623,7 +628,7 @@ export const ContentSuite: TestSuite = {
         assert.strictEqual(memberInfoB?.extension === `CPP`, true);
         assert.strictEqual(memberInfoB?.text === `C++ HEADER`, true);
 
-        const memberInfoC = await content?.getMemberInfo(`QSYSINC`, `H`, `OH_NONO` );
+        const memberInfoC = await component.getMemberInfo(`QSYSINC`, `H`, `OH_NONO` );
         assert.ok(!memberInfoC);
       }
     },
